+++ game_logic.py
@@ -1,3 +1,5 @@
+from Database.database import get_word_info, get_random_word, get_word_list
+
 ## TEST Global Var
 playerScore = 0
 playerRank = 0
@@ -38,6 +40,12 @@ def CheckValidity(guess):
     ## Word not valid
     return 1
 
+def GenerateRandomPuzzle():
+    random_puzzle = get_word_info()
+    current_puzzle = get_random_word(random_puzzle)
+    print(current_puzzle)
+    word_list = get_word_list(current_puzzle[0])
+    print(word_list)
 
 """
 CheckGuess take One Paramater
@@ -151,3 +159,4 @@ print (CheckValidity("abcd"))
 print (str(playerGuesses))
 print (playerScore)
  """
+GenerateRandomPuzzle()
